% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_acs_score.R
\name{calculate_acs_score}
\alias{calculate_acs_score}
\title{Calculate ACS Guideline Score}
\usage{
calculate_acs_score(df)
}
\arguments{
\item{df}{A \code{data.frame} with the required elements calculated by the
categorize functions and \code{calculate_acs_diet_score()}}
}
\value{
A \code{data.frame} with the subcomponent scores and the total score
as new columns:
\itemize{
\item \code{ACS_BMI} = The BMI subcomponent score
\item \code{ACS_MVPA} = The physical activity subcomponent score
\item \code{ACS_DIET} = The diet subcomponent score
\item \code{ACS_ALC} = The alcohol consumption subcomponent score
\item \code{ACS_TOTAL} = The sum of the components
}
}
\description{
This function creates the total American Cancer Society Guideline Score by
creating and combining the four subcomponents (BMI, MVPA, DIET, ALC)
}
\details{
<Add more details here about what is being done, columns that are required to be in the data, etc.>
}
\examples{
\dontrun{
score_df <- calculate_acs_score(df = my_score_data)
}

}
