% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/categorize_alc.R
\name{categorize_alc}
\alias{categorize_alc}
\title{Clean and categorize alcohol consumption}
\usage{
categorize_alc(ALC)
}
\arguments{
\item{ALC}{a numeric vector of drinks per day}
}
\value{
a numeric vector of categorized drinks per day:
\itemize{
\item \code{1} = No drinks
\item \code{2} = 1 drink per day or less
\item \code{3} = >1 to 2 drinks per day
\item \code{4} = More than 2 drinks per day
\item \code{9} = Missing
}
}
\description{
This function cleans alcohol intake variable (in number of drinks per day)
used in the creation of the alcohol subcomponent score in the total
American Cancer Society Guideline Score
}
\examples{
# Create sample vector
dpd <- c(
  floor(runif(n = 100, min = 0, max = 9)),
  rep(NA, 10)
)

alc_cat <- categorize_alc(dpd)
table(alc_cat)

}
