% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/categorize_bmi.R
\name{categorize_bmi}
\alias{categorize_bmi}
\title{Clean and categorize BMI at two time-points}
\usage{
categorize_bmi(BMI)
}
\arguments{
\item{BMI}{a numeric vector of BMI values}
}
\value{
a numeric vector of categorized BMI:
\itemize{
\item \code{1} = Normal weight (18.5-25 kg/m^2)
\item \code{2} = Overweight (25-<30 kg/m^2)
\item \code{3} = Obese (>=30 kg/m^2)
\item \code{9} = Missing
}
}
\description{
This function cleans body mass index variables from two time points
used in the creation of the BMI subcomponent score in the total
American Cancer Society Guideline Score
}
\examples{
# Create sample vector
bmi <- round(rnorm(n = 100, mean = 27, sd = 10), digits = 1)

bmi_cat <- categorize_bmi(bmi)
table(bmi_cat)

}
