% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/categorize_mvpa.R
\name{categorize_mvpa}
\alias{categorize_mvpa}
\title{Clean and categorize BMI at two time-points}
\usage{
categorize_mvpa(MVPA)
}
\arguments{
\item{MVPA}{a numeric vector of MVPA values}
}
\value{
a numeric vector of categorized MVPA:
\itemize{
\item \code{1} = None
\item \code{2} = >0 - <7.5 MET-hrs/week
\item \code{3} = 7.5 - <15.0 MET-hrs/week
\item \code{4} = >=15.0 MET-hrs/week
\item \code{9} = Missing
}
}
\description{
This function cleans moderate-vigorous physical activity variables
(in MET-hrs/week) used in the creation of the MVPA subcomponent score in the
total American Cancer Society Guideline Score
}
\examples{
# Create sample vector
mvpa <- c(
  round(rnorm(n = 100, mean = 8, sd = 3), digits = 1),
  rep(NA, 10)
)

mvpa_cat <- categorize_mvpa(mvpa)
table(mvpa_cat)

}
